% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, x0, y0, n = 10L, a = 1L, b = 2L)
}
\arguments{
\item{N}{the number of samples}

\item{x0}{initialization}

\item{y0}{initialization}

\item{n}{default is 10}

\item{a}{default is 1}

\item{b}{default is 2}
}
\value{
a random sample of size n
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
N <- 5000
x0 <- 0.5*n
y0 <- (x0+a) / (n+a+b)
Chain_Cpp <- gibbsC(N, x0, y0)
}
}
