% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{LDP}
\alias{LDP}
\title{Low Dimensional Projection Estimation}
\usage{
LDP(X, y, alpha = 0.05)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response}

\item{alpha}{the significance level alpha, default is 0.05}
}
\value{
A LDPE object is returned, including confidence intervals for regression coefficients.
}
\description{
Algorithms for fitting Low Dimensional Projection Estimation
}
\examples{
\dontrun{
dat <- simdata(n, p, rho, truebeta)
X <- dat$X
y <- dat$y
ldpe <- LDP(X, y)
CI <- ldpe$CI; len_CI <- ldpe$len_CI
}
}
