% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{simdata}
\alias{simdata}
\title{Simluate Toeplitz design and its response}
\usage{
simdata(n, p, rho, truebeta)
}
\arguments{
\item{n}{the datasize}

\item{p}{the number of variables}

\item{rho}{the correlation between variables}

\item{truebeta}{the regression coefficients}
}
\value{
Toeplitz design and its response
}
\description{
Simluate Toeplitz design and its response
}
\examples{
\dontrun{
n <- 200;  p <- 300;  rho <- 0.8; s <- 20
S_0 <- c(1:s)
truebeta <- rep(0, p)
truebeta[S_0] <- runif(s, 0, 2)
dat <- simdata(n, p, rho, truebeta)
}
}
