\name{WAP-package}
\alias{WAP-package}
\docType{package}
\title{
  Weighted Adaptive Projection
}
\description{
  Weighted adaptive projection (WAP) algorithm for parameter estimation in high dimensional logistic regression model. WAP reduces the amount of data required while guarantees parameter estimation accuracy by delicately distinguishing between identifiable and weak features.
}
\author{
  Hangjun Guo
  
  Maintainer: Hangjun Guo <guohangjun@cqu.edu.cn>
}
\keyword{ package }
\examples{
  ## see examples in WAP and LDP  
}
