% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{WAP}
\alias{WAP}
\title{Weighted Adaptive Projection}
\usage{
WAP(X, y, S, alpha = 0.05)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response}

\item{S}{the pre-screened identifiable feature set}

\item{alpha}{the significance level alpha, default is 0.05}
}
\value{
A WAP object is returned, including the estimated confidence intervals for LR model parameters.
}
\description{
Weighted Adaptive Projection
}
\examples{
\dontrun{
dat <- simdata(n, p, rho, truebeta)
X <- dat$X
y <- dat$y
S <- SIS(X, y, family = 'gaussian', tune = 'bic', seed = 2)$sis.ix0
wap <- WAP(X, y, S)
CI <- wap$CI; len_CI <- wap$len_CI
}
}
